requirements.txt
sanic==22.6.2
transformers
accelerate
openai-whisper @ git+https://github.com/Joemgu7/whisper-light.git
